services/rag_service/requirements.txt
grpcio>=1.76.0
grpcio-tools>=1.76.0
protobuf>=6.33.4
aio-pika>=9.0.0
qdrant-client>=1.7.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-huggingface
langchain-text-splitters
pypdf>=3.17.4
python-dotenv
python-docx
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
